SSCSSI-247: Added submit appeal v2 tests
