+++ build.gradle
@@ -350,7 +350,7 @@ dependencies {
     implementation group: 'com.mchange', name: 'c3p0', version: '0.10.1'
     implementation group: 'org.flywaydb', name: 'flyway-core', version: '9.22.3'
 
-    implementation group: 'com.github.hmcts', name: 'sscs-common', version: '5.12.11'
+    implementation group: 'com.github.hmcts', name: 'sscs-common', version: '3514bf1'
 
     implementation group: 'com.github.hmcts', name: 'sscs-pdf-email-common', version: '5.2.16'
