+++ src/test/java/uk/gov/hmcts/reform/sscs/service/AbstractSubmitAppealServiceTest.java
@@ -7,24 +7,14 @@ import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyMap;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.BDDMockito.given;
-import static org.mockito.BDDMockito.then;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static uk.gov.hmcts.reform.sscs.ccd.domain.EventType.ASSOCIATE_CASE;
-import static uk.gov.hmcts.reform.sscs.ccd.domain.EventType.DRAFT_TO_INCOMPLETE_APPLICATION;
-import static uk.gov.hmcts.reform.sscs.ccd.domain.EventType.DRAFT_TO_NON_COMPLIANT;
-import static uk.gov.hmcts.reform.sscs.ccd.domain.EventType.DRAFT_TO_VALID_APPEAL_CREATED;
-import static uk.gov.hmcts.reform.sscs.ccd.domain.EventType.INCOMPLETE_APPLICATION_RECEIVED;
-import static uk.gov.hmcts.reform.sscs.ccd.domain.EventType.NON_COMPLIANT;
-import static uk.gov.hmcts.reform.sscs.ccd.domain.EventType.SEND_TO_DWP;
-import static uk.gov.hmcts.reform.sscs.ccd.domain.EventType.VALID_APPEAL_CREATED;
 import static uk.gov.hmcts.reform.sscs.ccd.domain.State.READY_TO_LIST;
 import static uk.gov.hmcts.reform.sscs.util.SyaServiceHelper.getSyaCaseWrapper;
 
@@ -32,9 +22,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import feign.FeignException;
 import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
@@ -54,15 +42,10 @@ import org.mockito.junit.MockitoRule;
 import org.mockito.quality.Strictness;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.pdf.service.client.PDFServiceClient;
-import uk.gov.hmcts.reform.sscs.ccd.domain.Appeal;
 import uk.gov.hmcts.reform.sscs.ccd.domain.Appellant;
-import uk.gov.hmcts.reform.sscs.ccd.domain.BenefitType;
-import uk.gov.hmcts.reform.sscs.ccd.domain.Identity;
-import uk.gov.hmcts.reform.sscs.ccd.domain.MrnDetails;
 import uk.gov.hmcts.reform.sscs.ccd.domain.RegionalProcessingCenter;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseDetails;
-import uk.gov.hmcts.reform.sscs.ccd.exception.CcdException;
 import uk.gov.hmcts.reform.sscs.ccd.service.CcdService;
 import uk.gov.hmcts.reform.sscs.ccd.service.UpdateCcdCaseService;
 import uk.gov.hmcts.reform.sscs.config.CitizenCcdService;
@@ -72,7 +55,6 @@ import uk.gov.hmcts.reform.sscs.domain.wrapper.SyaCaseWrapper;
 import uk.gov.hmcts.reform.sscs.domain.wrapper.SyaContactDetails;
 import uk.gov.hmcts.reform.sscs.domain.wrapper.SyaMrn;
 import uk.gov.hmcts.reform.sscs.exception.ApplicationErrorException;
-import uk.gov.hmcts.reform.sscs.exception.DuplicateCaseException;
 import uk.gov.hmcts.reform.sscs.helper.EmailHelper;
 import uk.gov.hmcts.reform.sscs.idam.IdamService;
 import uk.gov.hmcts.reform.sscs.idam.IdamTokens;
@@ -91,7 +73,7 @@ public abstract class AbstractSubmitAppealServiceTest {
     public MockitoRule rule = MockitoJUnit.rule().strictness(Strictness.LENIENT);
 
     @Mock
-    private CcdService ccdService;
+    protected CcdService ccdService;
 
     @Mock
     private CitizenCcdService citizenCcdService;
@@ -115,23 +97,24 @@ public abstract class AbstractSubmitAppealServiceTest {
     private VenueService venueService;
 
     @Mock
-    private PDFServiceClient pdfServiceClient;
+    protected PDFServiceClient pdfServiceClient;
 
     @Mock
     private EmailHelper emailHelper;
 
     @Mock
-    private UpdateCcdCaseService updateCcdCaseService;
+    protected UpdateCcdCaseService updateCcdCaseService;
 
-    private uk.gov.hmcts.reform.sscs.service.SubmitAppealService submitAppealService;
-    private SubmitAppealService submitAppealServiceV2;
+    protected uk.gov.hmcts.reform.sscs.service.SubmitAppealService submitAppealService;
+    protected SubmitAppealService submitAppealServiceV2;
 
-    private final SyaCaseWrapper appealData = getSyaCaseWrapper();
+    protected final SyaCaseWrapper appealData = getSyaCaseWrapper();
 
-    private final String userToken = "user token";
+    protected final String userToken = "user token";
+
 
     @Captor
-    private ArgumentCaptor<SscsCaseData> capture;
+    protected ArgumentCaptor<SscsCaseData> capture;
 
     public static final String BIRMINGHAM_RPC = """
             {
@@ -218,207 +201,6 @@ public abstract class AbstractSubmitAppealServiceTest {
 
         given(idamService.getUserDetails(anyString())).willReturn(UserDetails.builder().roles(List.of("citizen")).build());
         given(emailHelper.generateUniqueEmailId(any(Appellant.class))).willReturn("Bloggs_33C");
-    }
-
-    @Test
-    public void givenCaseDoesNotExistInCcd_shouldCreateCaseWithAppealDetailsWithValidAppealCreatedEvent() {
-        byte[] expected = {};
-        given(pdfServiceClient.generateFromHtml(any(byte[].class), any())).willReturn(expected);
-
-        given(ccdService.findCcdCaseByNinoAndBenefitTypeAndMrnDate(anyString(), anyString(), anyString(), any())).willReturn(null);
-
-        submitAppealService.submitAppeal(appealData, userToken);
-
-        verify(ccdService).createCase(capture.capture(), eq(VALID_APPEAL_CREATED.getCcdType()), any(String.class), any(String.class), any(IdamTokens.class));
-        assertEquals("No", capture.getValue().getIsSaveAndReturn());
-    }
-
-    @Test
-    public void givenDraftCaseDoesExistAndCaseSubmitted_shouldUpdateCaseWithAppealDetailsWithDraftToValidAppealCreatedEvent() {
-        byte[] expected = {};
-        given(pdfServiceClient.generateFromHtml(any(byte[].class), any())).willReturn(expected);
-
-        given(ccdService.findCcdCaseByNinoAndBenefitTypeAndMrnDate(anyString(), anyString(), anyString(), any())).willReturn(null);
-        given(ccdService.getByCaseId(eq(123L), any())).willReturn(SscsCaseDetails.builder().build());
-
-        appealData.setCcdCaseId("123");
-        appealData.setIsSaveAndReturn("Yes");
-        submitAppealService.submitAppeal(appealData, userToken);
-
-        verify(ccdService).updateCase(capture.capture(), eq(123L), eq(DRAFT_TO_VALID_APPEAL_CREATED.getCcdType()), any(String.class), any(String.class), any(IdamTokens.class));
-        assertEquals("Yes", capture.getValue().getIsSaveAndReturn());
-
-    }
-
-    @Test
-    public void givenDraftCaseDoesExistAndCaseSubmittedHasNullBenefitType_shouldUpdateCaseWithAppealDetailsWithDraftToValidAppealCreatedEvent() {
-        byte[] expected = {};
-        given(pdfServiceClient.generateFromHtml(any(byte[].class), any())).willReturn(expected);
-
-        given(ccdService.findCaseBy(eq("data.appeal.appellant.identity.nino"), eq(appealData.getAppellant().getNino()), any()))
-                .willReturn(Collections.singletonList(
-                        SscsCaseDetails.builder()
-                                .id(12345678L)
-                                .data(SscsCaseData.builder()
-                                        .appeal(Appeal.builder()
-                                                .appellant(Appellant.builder().identity(Identity.builder().nino(appealData.getAppellant().getNino()).build()).build())
-                                                .mrnDetails(MrnDetails.builder().mrnDate(appealData.getMrn().getDate().plusDays(5).format(DateTimeFormatter.ISO_LOCAL_DATE))
-                                                        .build()).build()).build()).build()));
-        given(ccdService.findCcdCaseByNinoAndBenefitTypeAndMrnDate(anyString(), anyString(), anyString(), any())).willReturn(null);
-        given(ccdService.getByCaseId(eq(123L), any())).willReturn(SscsCaseDetails.builder().build());
-
-        appealData.setCcdCaseId("123");
-        appealData.setIsSaveAndReturn("Yes");
-        submitAppealService.submitAppeal(appealData, userToken);
-
-        verify(ccdService).updateCase(capture.capture(), eq(123L), eq(DRAFT_TO_VALID_APPEAL_CREATED.getCcdType()), any(String.class), any(String.class), any(IdamTokens.class));
-        assertEquals("Yes", capture.getValue().getIsSaveAndReturn());
-
-    }
-
-    @Test
-    public void givenDraftCaseDoesExistAndCaseSubmittedHasNullNino_shouldUpdateCaseWithAppealDetailsWithDraftToValidAppealCreatedEvent() {
-        byte[] expected = {};
-        given(pdfServiceClient.generateFromHtml(any(byte[].class), any())).willReturn(expected);
-
-        given(ccdService.findCaseBy(eq("data.appeal.appellant.identity.nino"), eq(appealData.getAppellant().getNino()), any()))
-                .willReturn(Collections.singletonList(
-                        SscsCaseDetails.builder()
-                                .id(12345678L)
-                                .data(SscsCaseData.builder()
-                                        .appeal(Appeal.builder()
-                                                .appellant(Appellant.builder().build())
-                                                .benefitType(BenefitType.builder().code(appealData.getBenefitType().getCode()).build())
-                                                .mrnDetails(MrnDetails.builder().mrnDate(appealData.getMrn().getDate().plusDays(5).format(DateTimeFormatter.ISO_LOCAL_DATE))
-                                                        .build()).build()).build()).build()));
-        given(ccdService.findCcdCaseByNinoAndBenefitTypeAndMrnDate(anyString(), anyString(), anyString(), any())).willReturn(null);
-        given(ccdService.getByCaseId(eq(123L), any())).willReturn(SscsCaseDetails.builder().build());
-
-        appealData.setCcdCaseId("123");
-        appealData.setIsSaveAndReturn("Yes");
-        submitAppealService.submitAppeal(appealData, userToken);
-
-        verify(ccdService).updateCase(capture.capture(), eq(123L), eq(DRAFT_TO_VALID_APPEAL_CREATED.getCcdType()), any(String.class), any(String.class), any(IdamTokens.class));
-        assertEquals("Yes", capture.getValue().getIsSaveAndReturn());
-
-    }
-
-    @Test
-    public void givenAssociatedCaseAlreadyExistsInCcd_shouldCreateCaseWithAppealDetailsAndAssociatedCase() {
-        byte[] expected = {};
-        given(pdfServiceClient.generateFromHtml(any(byte[].class), any())).willReturn(expected);
-
-        given(ccdService.findCaseBy(eq("data.appeal.appellant.identity.nino"), eq(appealData.getAppellant().getNino()), any()))
-                .willReturn(Collections.singletonList(
-                        SscsCaseDetails.builder()
-                                .id(12345678L)
-                                .data(SscsCaseData.builder()
-                                        .appeal(Appeal.builder()
-                                                .appellant(Appellant.builder().identity(Identity.builder().nino(appealData.getAppellant().getNino()).build()).build())
-                                                .benefitType(BenefitType.builder().code(appealData.getBenefitType().getCode()).build())
-                                                .mrnDetails(MrnDetails.builder().mrnDate(appealData.getMrn().getDate().plusDays(5).format(DateTimeFormatter.ISO_LOCAL_DATE))
-                                                        .build()).build()).build()).build()
-        ));
-
-        submitAppealService.submitAppeal(appealData, userToken);
-
-        verify(ccdService).createCase(capture.capture(), eq(VALID_APPEAL_CREATED.getCcdType()), any(String.class), any(String.class), any(IdamTokens.class));
-        assertEquals(1, capture.getValue().getAssociatedCase().size());
-        assertEquals("12345678", capture.getValue().getAssociatedCase().get(0).getValue().getCaseReference());
-    }
-
-    @Test
-    public void givenAssociatedCaseAlreadyExistsInCcd_shouldCreateCaseWithAppealDetailsAndAssociatedCaseWithoutMrn() {
-        byte[] expected = {};
-        given(pdfServiceClient.generateFromHtml(any(byte[].class), any())).willReturn(expected);
-
-        given(ccdService.findCaseBy(eq("data.appeal.appellant.identity.nino"), eq(appealData.getAppellant().getNino()), any()))
-                .willReturn(Collections.singletonList(
-                        SscsCaseDetails.builder()
-                                .id(12345678L)
-                                .data(SscsCaseData.builder()
-                                        .appeal(Appeal.builder()
-                                                .appellant(Appellant.builder().identity(Identity.builder().nino(appealData.getAppellant().getNino()).build()).build())
-                                                .benefitType(BenefitType.builder().code(appealData.getBenefitType().getCode()).build())
-                                                .mrnDetails(MrnDetails.builder().build())
-                                                .build()).build()).build()
-                ));
-
-        submitAppealService.submitAppeal(appealData, userToken);
-
-        verify(ccdService).createCase(capture.capture(), eq(VALID_APPEAL_CREATED.getCcdType()), any(String.class), any(String.class), any(IdamTokens.class));
-        assertEquals(1, capture.getValue().getAssociatedCase().size());
-        assertEquals("12345678", capture.getValue().getAssociatedCase().get(0).getValue().getCaseReference());
-    }
-
-    @Test
-    public void givenCaseDoesNotExistInCcdAndMrnDateIsMissing_shouldCreateCaseWithAppealDetailsWithIncompleteApplicationEvent() {
-        byte[] expected = {};
-        appealData.getMrn().setDate(null);
-
-        given(pdfServiceClient.generateFromHtml(any(byte[].class), any())).willReturn(expected);
-
-        given(ccdService.findCcdCaseByNinoAndBenefitTypeAndMrnDate(anyString(), anyString(), anyString(), any())).willReturn(null);
-
-        submitAppealService.submitAppeal(appealData, userToken);
-
-        verify(ccdService).createCase(capture.capture(), eq(INCOMPLETE_APPLICATION_RECEIVED.getCcdType()), any(String.class), any(String.class), any(IdamTokens.class));
-        verify(ccdService, times(0)).updateCase(any(SscsCaseData.class), eq(123L), eq(SEND_TO_DWP.getCcdType()), any(String.class), any(String.class), any(IdamTokens.class));
-        assertEquals("No", capture.getValue().getIsSaveAndReturn());
-    }
-
-    @Test
-    public void givenDraftCaseDoesExistAndMrnDateIsMissingAndCaseSubmitted_shouldUpdateCaseWithAppealDetailsWithDraftToIncompleteApplicationEvent() {
-        byte[] expected = {};
-        appealData.getMrn().setDate(null);
-
-        given(pdfServiceClient.generateFromHtml(any(byte[].class), any())).willReturn(expected);
-
-        given(ccdService.findCcdCaseByNinoAndBenefitTypeAndMrnDate(anyString(), anyString(), anyString(), any())).willReturn(null);
-
-        given(ccdService.getByCaseId(eq(123L), any())).willReturn(SscsCaseDetails.builder().build());
-
-        appealData.setCcdCaseId("123");
-        appealData.setIsSaveAndReturn("Yes");
-        submitAppealService.submitAppeal(appealData, userToken);
-
-        verify(ccdService).updateCase(capture.capture(), eq(123L), eq(DRAFT_TO_INCOMPLETE_APPLICATION.getCcdType()), any(String.class), any(String.class), any(IdamTokens.class));
-        assertEquals("Yes", capture.getValue().getIsSaveAndReturn());
-
-    }
-
-    @Test
-    public void givenCaseDoesNotExistInCcdAndMrnDateIsGreaterThan13Months_shouldCreateCaseWithAppealDetailsWithNonCompliantReceivedEvent() {
-        byte[] expected = {};
-        appealData.getMrn().setDate(LocalDate.now().minusMonths(13).minusDays(1));
-
-        given(pdfServiceClient.generateFromHtml(any(byte[].class), any())).willReturn(expected);
-
-        given(ccdService.findCcdCaseByNinoAndBenefitTypeAndMrnDate(anyString(), anyString(), anyString(), any())).willReturn(null);
-
-        submitAppealService.submitAppeal(appealData, userToken);
-
-        verify(ccdService).createCase(capture.capture(), eq(NON_COMPLIANT.getCcdType()), any(String.class), any(String.class), any(IdamTokens.class));
-        assertEquals("No", capture.getValue().getIsSaveAndReturn());
-    }
-
-    @Test
-    public void givenDraftCaseDoesExistAndMrnDateIsGreaterThan13MonthsAndCaseSubmitted_shouldUpdateCaseWithAppealDetailsWithDraftToNonCompliantEvent() {
-        byte[] expected = {};
-        appealData.getMrn().setDate(LocalDate.now().minusMonths(13).minusDays(1));
-
-        given(pdfServiceClient.generateFromHtml(any(byte[].class), any())).willReturn(expected);
-
-        given(ccdService.findCcdCaseByNinoAndBenefitTypeAndMrnDate(anyString(), anyString(), anyString(), any())).willReturn(null);
-
-        given(ccdService.getByCaseId(eq(123L), any())).willReturn(SscsCaseDetails.builder().build());
-
-        appealData.setCcdCaseId("123");
-        appealData.setIsSaveAndReturn("Yes");
-        submitAppealService.submitAppeal(appealData, userToken);
-
-        verify(ccdService).updateCase(capture.capture(), eq(123L), eq(DRAFT_TO_NON_COMPLIANT.getCcdType()), any(String.class), any(String.class), any(IdamTokens.class));
-        assertEquals("Yes", capture.getValue().getIsSaveAndReturn());
     }
 
     public abstract void givenSaveCaseWillReturnSaveCaseOperation(CitizenCcdService citizenCcdService,
@@ -824,84 +606,15 @@ public abstract class AbstractSubmitAppealServiceTest {
         assertEquals("UC Recovery from Estates", caseData.getAppeal().getMrnDetails().getDwpIssuingOffice());
     }
 
-    @Test(expected = CcdException.class)
-    public void givenExceptionWhenSearchingForCaseInCcd_shouldThrowException() {
-        given(ccdService.findCaseBy(eq("data.appeal.appellant.identity.nino"), eq(appealData.getAppellant().getNino()), any(IdamTokens.class)))
-            .willThrow(RuntimeException.class);
-
-        submitAppealService.submitAppeal(appealData, userToken);
-    }
-
-    @Test(expected = CcdException.class)
-    public void givenCaseDoesNotExistInCcdAndGivenExceptionWhenCreatingCaseInCcd_shouldThrowException() {
-        given(ccdService.findCcdCaseByNinoAndBenefitTypeAndMrnDate(anyString(), anyString(), anyString(), any(IdamTokens.class)))
-            .willReturn(null);
-
-        given(ccdService.createCase(any(SscsCaseData.class), any(String.class), any(String.class), any(String.class), any(IdamTokens.class)))
-            .willThrow(RuntimeException.class);
-
-        submitAppealService.submitAppeal(appealData, userToken);
-    }
-
-    @Test(expected = DuplicateCaseException.class)
-    public void givenCaseIsADuplicate_shouldNotResendEmails() {
-        given(ccdService.findCaseBy(eq("data.appeal.appellant.identity.nino"), eq(appealData.getAppellant().getNino()), any()))
-                .willReturn(Collections.singletonList(
-                        SscsCaseDetails.builder()
-                                .id(12345678L)
-                                .data(SscsCaseData.builder()
-                                        .appeal(Appeal.builder()
-                                                .appellant(Appellant.builder().identity(Identity.builder().nino(appealData.getAppellant().getNino()).build()).build())
-                                                .benefitType(BenefitType.builder().code(appealData.getBenefitType().getCode()).build())
-                                                .mrnDetails(MrnDetails.builder().mrnDate(appealData.getMrn().getDate().format(DateTimeFormatter.ISO_LOCAL_DATE))
-                                                        .build()).build()).build()).build()
-                ));
-
-        submitAppealService.submitAppeal(appealData, userToken);
-
-        then(pdfServiceClient).should(never()).generateFromHtml(any(byte[].class), anyMap());
-    }
-
-    @Test(expected = DuplicateCaseException.class)
-    public void givenCaseAlreadyExistsInCcd_shouldNotCreateCaseWithAppealDetails() {
-        given(ccdService.findCaseBy(eq("data.appeal.appellant.identity.nino"), eq(appealData.getAppellant().getNino()), any()))
-                .willReturn(Arrays.asList(
-                        SscsCaseDetails.builder()
-                                .data(SscsCaseData.builder()
-                                        .appeal(Appeal.builder()
-                                                .appellant(Appellant.builder().identity(Identity.builder().nino(appealData.getAppellant().getNino()).build()).build())
-                                                .benefitType(BenefitType.builder().code(appealData.getBenefitType().getCode()).build())
-                                                .mrnDetails(MrnDetails.builder().mrnDate(appealData.getMrn().getDate().format(DateTimeFormatter.ISO_LOCAL_DATE))
-                                                        .build()).build()).build()).build(),
-                        SscsCaseDetails.builder()
-                                .data(SscsCaseData.builder()
-                                        .appeal(Appeal.builder()
-                                                .appellant(Appellant.builder().identity(Identity.builder().nino(appealData.getAppellant().getNino()).build()).build())
-                                                .benefitType(BenefitType.builder().code("ESA").build())
-                                                .mrnDetails(MrnDetails.builder().mrnDate(appealData.getMrn().getDate().format(DateTimeFormatter.ISO_LOCAL_DATE))
-                                                        .build()).build()).build()).build(),
-                        SscsCaseDetails.builder()
-                                .data(SscsCaseData.builder()
-                                        .appeal(Appeal.builder()
-                                                .appellant(Appellant.builder().identity(Identity.builder().nino(appealData.getAppellant().getNino()).build()).build())
-                                                .benefitType(BenefitType.builder().code(appealData.getBenefitType().getCode()).build())
-                                                .mrnDetails(MrnDetails.builder().mrnDate(LocalDate.now().format(DateTimeFormatter.ISO_LOCAL_DATE))
-                                                        .build()).build()).build()).build()
-                ));
-
-        submitAppealService.submitAppeal(appealData, userToken);
-
-        verify(ccdService, never()).createCase(any(SscsCaseData.class), any(String.class), any(String.class), any(String.class), any(IdamTokens.class));
-    }
-
     @Test
     public void givenAssociatedCase_thenAddAssociatedCaseLinkToCase() {
         SscsCaseDetails matchingCase = SscsCaseDetails.builder().id(12345678L).data(SscsCaseData.builder().build()).build();
         List<SscsCaseDetails> matchedByNinoCases = new ArrayList<>();
         matchedByNinoCases.add(matchingCase);
 
-        SscsCaseData caseData = submitAppealService.addAssociatedCases(
-            SscsCaseData.builder().ccdCaseId("00000000").build(),
+        SscsCaseData caseData = SscsCaseData.builder().ccdCaseId("00000000").build();
+        submitAppealService.addAssociatedCases(
+                caseData,
             matchedByNinoCases);
 
         assertEquals(1, caseData.getAssociatedCase().size());
@@ -917,8 +630,9 @@ public abstract class AbstractSubmitAppealServiceTest {
         matchedByNinoCases.add(matchingCase1);
         matchedByNinoCases.add(matchingCase2);
 
-        SscsCaseData caseData = submitAppealService.addAssociatedCases(
-                SscsCaseData.builder().ccdCaseId("00000000").build(),
+        SscsCaseData caseData = SscsCaseData.builder().ccdCaseId("00000000").build();
+        submitAppealService.addAssociatedCases(
+                caseData,
                 matchedByNinoCases);
 
         assertEquals(2, caseData.getAssociatedCase().size());
@@ -931,8 +645,9 @@ public abstract class AbstractSubmitAppealServiceTest {
     public void addNoAssociatedCases() {
         List<SscsCaseDetails> matchedByNinoCases = new ArrayList<>();
 
-        SscsCaseData caseData = submitAppealService.addAssociatedCases(
-            SscsCaseData.builder().ccdCaseId("00000000").build(),
+        SscsCaseData caseData = SscsCaseData.builder().ccdCaseId("00000000").build();
+        submitAppealService.addAssociatedCases(
+                caseData,
             matchedByNinoCases);
 
         assertNull(caseData.getAssociatedCase());
